scripts/requirements-og.txt
# For scripts/generateOgImage.py
Pillow>=9.2.0
